% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assign_map}
\alias{assign_map}
\title{Assign edges to the hex grid}
\usage{
assign_map(proj, eccentricity = 0.9, report_progress = TRUE)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{eccentricity}{eccentricity of ellipses, defined as half the distance
between foci divided by the semi-major axis. \eqn{e = sqrt{1 - b^2/a^2}},
where \eqn{e} is the eccentricity, \eqn{a} is the length of the semi-major
axis, and \eqn{b} is the length of the semi-minor axis. Eccentricity ranges
between 0 (perfect circle) and 1 (straight line between foci).}

\item{report_progress}{if \code{TRUE} then a progress bar is printed to the
console during the permutation testing procedure.}
}
\description{
Given an RMAPI project with a hex map already created, determine
  which edges intersect each hex. Assumes an elliptical projection with the
  start- and end-points of an edge becoming the two foci of an ellipse.
}
