% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{barrier_lines}
\alias{barrier_lines}
\title{Add lines to a plot representing barriers}
\usage{
barrier_lines(barrier_x = 0, barrier_y = 0, barrier_angle = 0, ...)
}
\arguments{
\item{barrier_x}{TODO}

\item{barrier_y}{TODO}

\item{barrier_angle}{TODO}

\item{...}{other arguments that get passed directly to \code{abline()}}
}
\description{
TODO - some help text here.
}
