% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Load data into RMAPI project}
\usage{
bind_data(proj, data, check_delete_output = TRUE)
}
\arguments{
\item{proj}{the current RMAPI project}

\item{data}{a data frame, formatted into the correct RMAPI format (see Details)}

\item{check_delete_output}{whether to perform a check to see if project already contains data, in which case all old data and output will be lost}
}
\description{
TODO - some help text here.
}
