% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_hex_values}
\alias{calc_hex_values}
\title{Partial version of rmapi_analysis - calculate hex values}
\usage{
calc_hex_values(
  proj,
  null_method = 1,
  n_perms = 100,
  n_breaks = 1,
  dist_breaks = NULL,
  empirical_tail = "both",
  report_progress = TRUE,
  min_intersections = 5
)
}
\description{
<<<<<<< HEAD
Partial version of rmapi_analysis - calculate hex values from pairwise data and previously calculated intersection and weighting data
=======
Partial version of rmapi_analysis - calculate hex values from
  pairwise data and previously calculated intersection and weighting data
>>>>>>> b3306fe8a3af59f022f32a93132fa5e22330e68a
}
