% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_intersections}
\alias{calc_intersections}
<<<<<<< HEAD
\title{Partial version of rmapi_analysis - calculate intersection and weighting data for making multiple maps}
=======
\title{Partial version of rmapi_analysis - calculate intersection and
  weighting data for making multiple maps}
>>>>>>> b3306fe8a3af59f022f32a93132fa5e22330e68a
\usage{
calc_intersections(proj, eccentricity = 0.5, min_intersections = 5)
}
\description{
<<<<<<< HEAD
Partial version of rmapi_analysis - calculate intersection and weighting data for making multiple maps
=======
Partial version of rmapi_analysis - calculate intersection and
  weighting data for making multiple maps.
>>>>>>> b3306fe8a3af59f022f32a93132fa5e22330e68a
}
