% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{check_RMAPI_loaded}
\alias{check_RMAPI_loaded}
\title{Check that RMAPI package has loaded successfully}
\usage{
check_RMAPI_loaded()
}
\description{
Simple function to check that RMAPI package has loaded
  successfully. Prints "RMAPI loaded successfully!" if so.
}
