% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords_DRC}
\alias{coords_DRC}
\title{example data set to test RMAPI functionality}
\format{An object of class \code{data.frame} with 529 rows and 2 columns.}
\usage{
data(coords_DRC)
}
\description{
Made up data for testing - this data set should be deleted prior to release.
\cr
\cr
Spatial points corresponding to real DRC demographic and health survey cluster locations. These points can be used to generate simulated data under different spatial barriers using the \code{sim_custom()} function.
}
\examples{
# TODO

}
\keyword{datasets}
