% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map2.R
\name{create_map2}
\alias{create_map2}
\title{Create mapping space}
\usage{
create_map2(proj, hex_size = 1, buffer = 0, outline)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{hex_size}{size of hexagons.}

\item{buffer}{size of buffer zone around the data. It is recommended to not
use a buffer to avoid edge-effects.}

\item{outline}{List of coordinates defining outline}
}
\description{
Create mapping space - alternate version where hex map is built around input region outline
}
