% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fakeData_DRC}
\alias{fakeData_DRC}
\title{example data set to test RMAPI functionality}
\format{An object of class \code{data.frame} with 529 rows and 532 columns.}
\usage{
data(fakeData_DRC)
}
\description{
Made up data for testing - this data set should be deleted prior to release.
\cr
\cr
Spatial points correspond to real DRC DHS cluster locations, giving some spatial realism. Values between points are completely made up, and represent a barrier to gene flow transecting the country from North to South. Hopefully RMAPI should be able to detect this.
}
\examples{
# create project and load in data
data(fakeData_DRC)
proj <- rmapi_project()
proj <- loadData(proj, fakeData_DRC)

# define parameters
proj$alpha <- 2

# run simulations
proj <- runSims(proj)

# plot output
RMAPI_plot1(proj)

}
\keyword{datasets}
