% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_model}
\alias{fit_model}
\title{Fit a simple model to data}
\usage{
fit_model(proj, type = 1)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{type}{which model to fit to the data: 1 = linear model, 2 =
exponential model.}
}
\description{
Fit a simple linear or exponential model to the distribution of
  the observed statistic as a function of distance. Can be used in the
  simulation step to generate a null model against which to compare.
}
