% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_model2}
\alias{fit_model2}
\title{Fit a simple model to data (alt)}
\usage{
fit_model2(
  proj,
  type = 1,
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  a_fit = 1,
  b_fit = 1,
  c_fit = 0
)
}
\description{
Alternate version of fit_model using x and y min and max; 1: Linear model, 2: SSasymp
model, 3: Power model, 4: Exponential model; a_fit, b_fit and c_fit are starting parameter values.
}
