% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_ellipse}
\alias{get_ellipse}
\title{Calculate ellipse polygon coordinates from foci and eccentricity}
\usage{
get_ellipse(f1 = c(-3, -2), f2 = c(3, 2), ecc = 0.8, n = 100)
}
\arguments{
\item{f1}{x- and y-coordinates of the first focus.}

\item{f2}{x- and y-coordinates of the first focus.}

\item{ecc}{eccentricity of the ellipse, defined as half the distance between
foci divided by the semi-major axis. We can say \eqn{e = sqrt{1 -
b^2/a^2}}, where \eqn{e} is the eccentricity, \eqn{a} is the length of the
semi-major axis, and \eqn{b} is the length of the semi-minor axis.
Eccentricity ranges between 0 (perfect circle) and 1 (straight line between
foci).}

\item{n}{number of points in polygon.}
}
\description{
Calculate ellipse polygon coordinates from foci and eccentricity.
}
