% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_significant_hexes}
\alias{get_significant_hexes}
\title{Get significant hexes}
\usage{
get_significant_hexes(
  proj,
  empirical_tail = "both",
  alpha_raw = 0.05,
  min_hex_coverage = 10
)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{empirical_tail}{whether to do calculate empirical p-values using a
one-sided test (\code{empirical_tail = "left"} or \code{empirical_tail =
"right"}) or a two-sided test (\code{empirical_tail = "both"}).}

\item{alpha_raw}{the significance threshold used to determine significantly
high/low values. This raw value is Bonferroni corrected based on the
effective number of independent samples, and hence applies to the whole map
and not just a single hex.}

\item{min_hex_coverage}{minimum coverage (number of edges assigned to a hex)
for it to be included in the final result.}
}
\description{
Given a completed RMAPI analysis, return the number of hexes
  that pass a stated significance threshold. Calculation takes account of the
  inherent correlation between hexes through the effective sample size
  calculation.
}
