% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_spatial_distance}
\alias{get_spatial_distance}
\title{Get great circle distance between spatial points}
\usage{
get_spatial_distance(long, lat)
}
\arguments{
\item{long, lat}{vector of longitudes and latitudes.}
}
\description{
Get great circle distance between spatial points, defined by a
  vector of longitudes and latitudes. Distances are returned in a pairwise
  distance matrix.
}
