% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapi_sim.R
\name{is.rmapi_sim}
\alias{is.rmapi_sim}
\title{Determine if object is of class rmapi_sim}
\usage{
is.rmapi_sim(x)
}
\arguments{
\item{x}{object to query.}
}
\description{
Determine if object is of class \code{rmapi_sim}.
}
