% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_coords}
\alias{load_coords}
\title{Load coordinates of sampling locations into RMAPI project}
\usage{
load_coords(proj, long, lat, check_delete_output = TRUE)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{long, lat}{vectors of sampling longitudes and latitudes.}

\item{check_delete_output}{if \code{TRUE} (the default) then check before
overwriting any existing data loaded into a project.}
}
\description{
Given an existing RMAPI project, load the spatial coordinates of
  points at which samples were obtained.
}
