% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_data}
\alias{load_data}
\title{Load pairwise data into RMAPI project}
\usage{
load_data(proj, pairwise_data, check_delete_output = TRUE)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{pairwise_data}{matrix of pairwise statistics between nodes.}

\item{check_delete_output}{if \code{TRUE} (the default) then check before
overwriting any existing data loaded into a project.}
}
\description{
Load pairwise statistical data into an existing RMAPI project.
}
