% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_old_data2}
\alias{load_old_data2}
\title{Load "simulation" data from original MAPI paper with added 'void'}
\usage{
load_old_data2()
}
\description{
Load "simulation" data from original MAPI paper with added 'void'
}
