% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_old_data4}
\alias{load_old_data4}
\title{Load "simulation" data from original MAPI paper with added 'voids'}
\usage{
load_old_data4()
}
\description{
Load "simulation" data from original MAPI paper with added 'voids'
}
