% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{loadData}
\alias{loadData}
\title{Load data into RMAPI project}
\usage{
loadData(proj, data, checkDeleteOutput = TRUE)
}
\arguments{
\item{proj}{the current RMAPI project}

\item{data}{a data frame, formatted into the correct RMAPI format (see Details)}

\item{checkDeleteOutput}{whether to perform a check to see if project already contains data, in which case all old data and output will be lost}
}
\description{
TODO - some help text here.
}
