% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{lonlat_to_bearing}
\alias{lonlat_to_bearing}
\title{Calculate great circle distance and bearing between coordinates}
\usage{
lonlat_to_bearing(origin_lon, origin_lat, dest_lon, dest_lat)
}
\arguments{
\item{origin_lon, origin_lat}{the origin longitude and latitude.}

\item{dest_lon, dest_lat}{the destination longitude and latitude}
}
\description{
Calculate great circle distance and bearing between spatial
  coordinates, defined by longitude and latitude of both origin and
  destination points.
}
\examples{
# one degree longitude should equal approximately 111km at the equator
lonlat_to_bearing(0, 0, 1, 0)
}
