% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapi_project.R
\name{plot.rmapi_project}
\alias{plot.rmapi_project}
\title{Overload plot function for RMAPI project}
\usage{
plot.rmapi_project(x, y, ...)
}
\arguments{
\item{x}{object of class \code{rmapi_project}.}

\item{y}{(ignored).}

\item{...}{(ignored).}
}
\description{
Overload plot function for RMAPI project.
}
