% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot hex coverage for a given eccentricity}
\usage{
plot_coverage(proj, eccentricity = 0.9, n_ell = 20, return_type = 1)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{eccentricity}{eccentricity of ellipses, defined as half the distance
between foci divided by the semi-major axis. We can say \eqn{e = sqrt{1 -
b^2/a^2}}, where \eqn{e} is the eccentricity, \eqn{a} is the length of the
semi-major axis, and \eqn{b} is the length of the semi-minor axis.
Eccentricity ranges between 0 (perfect circle) and 1 (straight line between
foci).}

\item{n_ell}{number of points that make up an ellipse.}

\item{return_type}{format of return object: 1 = a \code{gridExtra} object, 2
= a list of \code{ggplot2} objects.}
}
\description{
Plot hex coverage (number of intersecting ellipses) for a given
  eccentricity.
}
