% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_daily_states}
\alias{plot_daily_states}
\title{Plot daily counts of each host state from simulation}
\usage{
plot_daily_states(x, deme = 1, states = c("S", "E", "I"))
}
\arguments{
\item{x}{object of class \code{rmapi_sim}.}

\item{deme}{which deme to plot.}

\item{states}{which states to plot. Can be any subset of \code{c("S", "E",
"I", "Sv", "Ev", "Iv")}.}
}
\description{
For a set of simulation output produced by the function
  \code{sim_falciparum()}, plots daily simulated values in a given deme.
}
