% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Interactive hex map of RMAPI output}
\usage{
plot_leaflet(proj, variable = NULL, fill_opacity = 1,
  legend_opacity = 1)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{variable}{which element of the project output to use as map colours.}

\item{fill_opacity}{opacity of fill; 0 = fully transparent, 1 = fully opaque.}

\item{legend_opacity}{opacity of legend; 0 = fully transparent, 1 = fully
opaque.}
}
\description{
Interactive hex map of RMAPI output.
}
