% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map2}
\alias{plot_map2}
\title{Plot hex map of RMAPI output with hexes considered statistically significant highlighted}
\usage{
plot_map2(proj, col_scale = magma(100), barrier_list = list(), tails = c(0, 1))
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{col_scale}{the colour scale to use.}

\item{barrier_list}{optional list of polygon coordinates that are added to
plot.}

\item{variable}{which element of the project output to use as map colours.}
}
\description{
Plot hex map of RMAPI output.
}
