% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapi_project.R
\name{print.rmapi_project}
\alias{print.rmapi_project}
\title{Overload print function for RMAPI project}
\usage{
\method{print}{rmapi_project}(x, ...)
}
\arguments{
\item{x}{object of class \code{rmapi_project}.}

\item{...}{(ignored).}
}
\description{
Overload print function for RMAPI project.
}
