% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{print_full}
\alias{print_full}
\title{Print unclassed object}
\usage{
print_full(x)
}
\arguments{
\item{x}{object to print in full.}
}
\description{
Print object after unclassing, thereby removing any custom print
  method.
}
