% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{RMAPI_plot1}
\alias{RMAPI_plot1}
\title{Simple filled contour plot of RMAPI output}
\usage{
RMAPI_plot1(proj)
}
\arguments{
\item{proj}{the current RMAPI project}
}
\description{
TODO - some help text here.
}
