% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapi_project.R
\name{rmapi_project}
\alias{rmapi_project}
\title{Define empty RMAPI project}
\usage{
rmapi_project()
}
\description{
Define empty RMAPI project.
}
\details{
#------------------------------------------------
}
