% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainFunctions.R
\name{runSims}
\alias{runSims}
\title{Perform RMAPI simulation}
\usage{
runSims(proj, Nperms = 100)
}
\arguments{
\item{proj}{the current RMAPI project}
}
\description{
TODO - some help text here.
}
\examples{
runSims()
}
