% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_custom}
\alias{sim_custom}
\title{Simulate rectangular lattice of points with one or more barriers}
\usage{
sim_custom(x, y, barrier_x = 0, barrier_y = 0, barrier_angle = 0,
  barrier_penalty = 10)
}
\arguments{
\item{x}{positions of points in the x-dimension}

\item{y}{positions of points in the y-dimension}

\item{barrier_x}{TODO}

\item{barrier_y}{TODO}

\item{barrier_angle}{TODO}

\item{barrier_penalty}{TODO}
}
\description{
TODO - some help text here.
}
