% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_custom_barrier}
\alias{sim_custom_barrier}
\title{Create new distance-based pairwise data using input node positions with custom barrier}
\usage{
sim_custom_barrier(x, y, xbarrier, ybarrier, rbarrier, vbarrier,
  distance_model)
}
\description{
TODO - some help text here.
}
