% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_falciparum}
\alias{sim_falciparum}
\title{Simulate genetic data from simple P.falciparum dynamic model}
\usage{
sim_falciparum(a = 0.3, p = 0.85, mu = -log(p), u = 12, v = 10,
  g = 12, prob_infection = 0.6, duration_infection = dgeom(1:25,
  1/5), infectivity = 0.07, max_innoculations = 5, H = 1000,
  seed_infections = 100, M = 1000, time_out = 100)
}
\arguments{
\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on
humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. mu = -log(p) unless otherwise
specified.}

\item{u}{intrinsic incubation period. The number of days from infection to
blood-stage infection in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to
becoming infectious in a mosquito.}

\item{g}{lag time between human blood-stage infection and production of
gametocytes.}

\item{prob_infection}{probability a human becomes infected after being bitten
by an infected mosquito.}

\item{duration_infection}{vector specifying probability distribution of time
(in days) of a malaria episode.}

\item{infectivity}{probability a mosquito becomes infected after biting an
infective human host.}

\item{max_innoculations}{maximum number of innoculations that an individual
can hold simultaneously.}

\item{H}{human population size, which is assumed the same in every deme.}

\item{seed_infections}{vector specifying the initial number of infected
humans in each deme.}

\item{M}{vector specifying mosquito population size (strictly the number of
adult female mosquitoes) in each deme.}

\item{time_out}{vector of times (days) at which output is produced.}
}
\description{
Simulate genetic data from simple P.falciparum dynamic model.
}
