% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_rect}
\alias{sim_rect}
\title{Simulate rectangular lattice of points with linear barrier}
\usage{
sim_rect(Nx = 10, Ny = 10, sep = 1, barrier_x = 0, barrier_y = 0,
  barrier_angle = 0, barrier_penalty = 10)
}
\arguments{
\item{Nx}{number of points in the x-dimension}

\item{Ny}{number of points in the y-dimension}

\item{sep}{TODO}

\item{barrier_x}{TODO}

\item{barrier_y}{TODO}

\item{barrier_angle}{TODO}

\item{barrier_penalty}{TODO}
}
\description{
TODO - some help text here.
}
