% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_simple}
\alias{sim_simple}
\title{Simulate data from simple model}
\usage{
sim_simple(node_long, node_lat, barrier_list = list(),
  barrier_penalty = numeric(), barrier_method = 1, ecc = 0.9,
  n_ell = 20, dist_transform = 1, lambda = 0.1, eps = 0.1)
}
\arguments{
\item{node_long}{longitudes of nodes.}

\item{node_lat}{latitudes of nodes.}

\item{barrier_list}{list of polygons representing barriers. Each element of
the list must be a dataframe with columns \code{long} and \code{lat}
specifying the coordinates of points that make up the polygon. Polygons
must be complete rings, meaning the final row of the dataframe must equal
the first row.}

\item{barrier_penalty}{penalty values of each barrier.}

\item{barrier_method}{the method by which penalties are applied:
\itemize{
  \item{bullet 1 compare line, apply penalty on intersection}
  \item{bullet 2 compare line, apply penalty per unit intersection}
  \item{bullet 3 compare ellipse, apply penalty per unit area intersection}
}}

\item{ecc}{eccentricity of ellipses (only used under \code{barrier_method = 3}).}

\item{n_ell}{number of points that make up an ellipse (only used under \code{barrier_method = 3}).}

\item{dist_transform}{the method by which distances are transformed to produce the final statistic:
\itemize{
  \item{bullet 1 linear}
  \item{bullet 2 exponential decay}
}}

\item{lambda}{the rate of decay of the exponential function (only used under \code{dist_transform = 2}).}

\item{eps}{the standard deviation of white noise applied to final statistics.}
}
\description{
Simulate data from simple model.
}
