% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmapi_project.R
\name{summary.rmapi_project}
\alias{summary.rmapi_project}
\title{Overload summary function for RMAPI project}
\usage{
summary.rmapi_project(object, ...)
}
\arguments{
\item{object}{object of class \code{rmapi_project}.}

\item{...}{(ignored).}
}
\description{
Overload summary function for RMAPI project.
}
