% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Load data into RMAPI project}
\usage{
bind_data(proj, long, lat, stat_dist, check_delete_output = TRUE)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{long, lat}{vectors of node longitudes and latitudes.}

\item{stat_dist}{matrix of pairwise statistics between nodes.}

\item{check_delete_output}{if \code{TRUE} (the default) then check before
overwriting any existing data loaded into a project.}
}
\description{
Load data into an existing RMAPI project.
}
