% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bind_data}
\alias{bind_data}
\title{Load data into RMAPI project}
\usage{
bind_data(proj, long, lat, pairwise_dist, check_delete_output = TRUE)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{long}{vector of node longitudes.}

\item{lat}{vector of node latitudes.}

\item{pairwise_dist}{matrix of pairwise distances between nodes.}

\item{check_delete_output}{if \code{TRUE} (the default) then check before
overwriting any existing data loaded into a project.}
}
\description{
Load data into RMAPI project.
}
