% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_intersections}
\alias{calc_intersections}
\title{Partial version of rmapi_analysis - calculate intersection and
  weighting data for making multiple maps}
\usage{
calc_intersections(proj, eccentricity = 0.5, min_intersections = 5)
}
\description{
Partial version of rmapi_analysis - calculate intersection and
  weighting data for making multiple maps.
}
