% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_map}
\alias{create_map}
\title{Create map composed of hex tiles}
\usage{
create_map(proj, hex_size = NULL)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{hex_size}{size of hexagons.}
}
\description{
Given an RMAPI project with sampling coordinates already loaded,
  create a hex grid around these points.
}
