% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_map}
\alias{create_map}
\title{Create mapping space}
\usage{
create_map(proj, hex_size = 1, buffer = 0)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{hex_size}{size of hexagons.}

\item{buffer}{size of buffer zone around the data. It is recommended to not
use a buffer to avoid edge-effects.}
}
\description{
Create mapping space.
}
