% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_model}
\alias{fit_model}
\title{Fit a simple model to pairwise data}
\usage{
fit_model(
  proj,
  type = 1,
  x_min = -Inf,
  x_max = Inf,
  y_min = -Inf,
  y_max = Inf,
  a_init = 1,
  b_init = 1,
  c_init = 1
)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{type}{which model to fit to the data: 1 = linear model, 2 = asymptotic
model, 3 = power model.}

\item{x_min, x_max, y_min, y_max}{can be used to mask out values outside a
specified range.}

\item{a_init, b_init, c_init}{initial values used in model fitting for models 1
and 3.}
}
\description{
Using data already loaded into an RMAPI project, fits a simple
  model of pairwise statistical distances against spatial distances. This
  model can be used in the simulation step to generate a null expectation
  against which to compare. The available models are:
  \enumerate{
    \item Linear model, \code{y = a*x + b}
    \item Asymptotic model, \code{y = alpha + (beta - alpha)*exp(-exp(log_lambda)*x)}
    \item Power model, code{y = a*x^b + c}
  }
}
