% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_spatial_distance}
\alias{get_spatial_distance}
\title{Get great circle distance between spatial points}
\usage{
get_spatial_distance(long, lat)
}
\arguments{
\item{long}{vector of longitudes.}

\item{lat}{vector of latitudes.}
}
\description{
Get great circle distance between spatial points.
}
