% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot hex coverage}
\usage{
plot_coverage(proj, return_type = 1)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{return_type}{format of return object: 1 = a \code{gridExtra} object, 2
= a list of \code{ggplot2} objects.}
}
\description{
Given an RMAPI project with a hap already assigned, plot the
  coverage (number of intersecting ellipses) of each hex. Useful as a
  diagnostic plot, as hexes with low coverage will conform less well to the
  assumptions of the permutation test (see \emph{detials}).
}
\details{
Good coverage is needed to ensure the validity of the statistical
  procedure. For each hex, the observed value is the mean of the (normalised)
  values of the edges that intersect it. Under the null model that these
  normalised values have no systematic bias, the distribution of this test
  statistic is approximately normal as a consequence of the central limit
  theorem. Hence, we can use a permutation test to characterise the mean and
  standard deviation of this null distribution, then we can quantify how
  extreme the observed value is in terms of its z-score. If coverage is too
  low then the null distribution will not be normally distributed, and hence
  the z-score will not be an accurate description of how extreme the observed
  data really is.
}
