% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot pairwise spatial distance against loaded statistic}
\usage{
plot_dist(proj, col = "#00000050", overlay_model = TRUE)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{col}{the colour of points.}

\item{overlay_model}{whether to overlay the model fit, if \code{fit_model()}
has been called.}
}
\description{
Plot pairwise spatial distance against loaded statistic.
}
