% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Interactive hex map of RMAPI output}
\usage{
plot_leaflet(proj, variable = NULL, fill_opacity = 0.8,
  legend_opacity = 1, map_type = 97, col_scale = magma(100))
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{variable}{which element of the project output to use as map colours.}

\item{fill_opacity}{opacity of fill; 0 = fully transparent, 1 = fully opaque.}

\item{legend_opacity}{opacity of legend; 0 = fully transparent, 1 = fully
opaque.}

\item{map_type}{an index from 1 to 137 indicating the type of base map. The
map types are taken from \code{leaflet::providers}, see
\href{http://leaflet-extras.github.io/leaflet-providers/preview/index.html}{here}
for an interactive gallary.}

\item{col_scale}{the colour scale to use.}
}
\description{
Interactive hex map of RMAPI output.
}
