% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Plot hex map of RMAPI output}
\usage{
plot_map(
  proj,
  col_scale = viridisLite::magma(100),
  plot_sampling_points = TRUE,
  plot_hex_grid = TRUE,
  plot_significance = TRUE,
  barrier_list = list()
)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{col_scale}{the colour scale to use.}

\item{plot_sampling_points}{whether to overlay sampling locations.}

\item{plot_hex_grid}{whether to plot the hex map. If the project contains
output then hexes will be coloured based on z-scores, otherwise a flat
colour scheme will be used.}

\item{plot_significance}{whether to outline areas that were identified as
significant outliers.}

\item{barrier_list}{optional list of polygon coordinates that are added to
plot.}
}
\description{
Plot hex map of RMAPI output.
}
