% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Simple filled contour plot of RMAPI output}
\usage{
plot_map(proj, variable = NULL)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{variable}{which element of the project output to use as map colours.}
}
\description{
Simple filled contour plot of RMAPI output.
}
