% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rmapi_analysis}
\alias{rmapi_analysis}
\title{Perform RMAPI analysis}
\usage{
rmapi_analysis(
  proj,
  n_perms = 1000,
  n_breaks = 50,
  min_dist = 0,
  max_dist = Inf,
  min_group_size = 5,
  report_progress = TRUE
)
}
\arguments{
\item{proj}{object of class \code{rmapi_project}.}

\item{n_perms}{number of permutations in test.}

\item{n_breaks}{values are broken into this many groups based on spatial
distances. Permutation testing then occurs within groups. Hence, a larger
value of \code{n_breaks} does a better job of conditioning on spatial
distance, but comes at the cost of statistical power in permutation
testing. A warning is printed if any group ends up containing fewer than 10
edges.}

\item{min_dist, max_dist}{minimum and maximum edge lengths to be included in
the analysis. Anything outside this range is ignored.}

\item{min_group_size}{minimum number of edges within a spatial permutation
group, otherwise edges within this group are replaced with \code{NA}.}

\item{report_progress}{if \code{TRUE} then a progress bar is printed to the
console during the permutation testing procedure.}
}
\description{
TODO.
}
