% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_falciparum}
\alias{sim_falciparum}
\title{Simulate genetic data from simple P.falciparum dynamic model}
\usage{
sim_falciparum(L = 24, prob_cotransmission = 0.5, a = 0.3, p = 0.9,
  mu = -log(p), u = 12, v = 10, g = 10, prob_infection = seq(0.1,
  0.01, -0.01), duration_infection = dgeom(1:300, 1/50),
  infectivity = 1, max_innoculations = 5, H = 1000,
  seed_infections = 100, M = 1000, mig_matrix = diag(length(M)),
  time_out = 100, report_progress = TRUE)
}
\arguments{
\item{L}{number of loci. The maximum number of loci is 1000, as at higher
numbers haplotypes begin to exceed integer representation (2^L).}

\item{prob_cotransmission}{probability of mosquito transmitting multiple
haplotypes to host.}

\item{a}{human blood feeding rate. The proportion of mosquitoes that feed on
humans each day.}

\item{p}{mosquito probability of surviving one day.}

\item{mu}{mosquito instantaneous death rate. mu = -log(p) unless otherwise
specified.}

\item{u}{intrinsic incubation period. The number of days from infection to
blood-stage infection in a human host.}

\item{v}{extrinsic incubation period. The number of days from infection to
becoming infectious in a mosquito.}

\item{g}{lag time between human blood-stage infection and production of
gametocytes.}

\item{prob_infection}{probability a human becomes infected after being bitten
by an infected mosquito.}

\item{duration_infection}{vector specifying probability distribution of time
(in days) of a malaria episode.}

\item{infectivity}{probability a mosquito becomes infected after biting an
infective human host.}

\item{max_innoculations}{maximum number of innoculations that an individual
can hold simultaneously.}

\item{H}{human population size, which is assumed the same in every deme.}

\item{seed_infections}{vector specifying the initial number of infected
humans in each deme.}

\item{M}{vector specifying mosquito population size (strictly the number of
adult female mosquitoes) in each deme.}

\item{mig_matrix}{migration matrix specifing the daily probability of
migrating from each deme to each other deme. Migration must be equal in
both directions, meaning this matrix must be symmetric.}

\item{time_out}{vector of times (days) at which output is produced.}

\item{report_progress}{if \code{TRUE} then a progress bar is printed to the
console during simuation.}
}
\description{
Simulate genetic data from simple P.falciparum dynamic model.
}
